+++ parser.hpp
@@ -1,15 +1,17 @@
 #ifndef PARSER_HPP
 # define PARSER_HPP
 
+# include <string>
+
 typedef struct			s_state
 {
-	uint8_t				*board;
+	int				*board;
 	int					g;
 	int					h;
 	int					f;
 	struct s_state*		predecessor;
 }						t_state;
 
-uint8_t		*parse(std::string filename, size_t *size);
+int		*parse(std::string filename, size_t *size);
 
 #endif
